feat: 모임 참가 이력 등록 기능
- 모임 참가 이력 등록
- 모임 생성 시 주최자도 모임 참가자로 등록 (ACCEPTED 상태)
closes #42

+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/infrastructure/repository/JpaMeetingParticipationHistoryRepository.java
@@ -4,6 +4,6 @@ import com.learningcrew.linkup.meeting.command.domain.aggregate.MeetingParticipa
 import com.learningcrew.linkup.meeting.command.domain.repository.MeetingParticipationHistoryRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaMeetingParticipationHistoryRepository extends MeetingParticipationHistoryRepository, JpaRepository<MeetingParticipationHistory, Integer> {
+public interface JpaMeetingParticipationHistoryRepository extends MeetingParticipationHistoryRepository, JpaRepository<MeetingParticipationHistory, Long> {
 
 }
